+++ packages/metadata/tsconfig.json
@@ -1,8 +1,3 @@
 {
-  "extends": "../../tsconfig.esm.json",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./dist"
-  },
-  "include": ["./src"]
+	"extends": "../../tsconfig.json"
 }
