+++ packages/reorg/package.json
@@ -1,28 +1,28 @@
 {
-  "name": "@orgajs/reorg",
-  "version": "4.2.0",
-  "description": "orga processor for unifiedjs",
-  "files": [
-    "index.js",
-    "index.d.ts"
-  ],
-  "main": "index.js",
-  "exports": "./index.js",
-  "type": "module",
-  "author": "Xiaoxing Hu <hi@xiaoxing.dev>",
-  "license": "MIT",
-  "homepage": "https://github.com/orgapp/orgajs/tree/main/packages/reorg#readme",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/orgapp/orgajs.git",
-    "directory": "packages/reorg"
-  },
-  "scripts": {},
-  "dependencies": {
-    "@orgajs/reorg-parse": "workspace:^",
-    "unified": "^11.0.5"
-  },
-  "devDependencies": {
-    "@types/unist": "^3.0.3"
-  }
+	"name": "@orgajs/reorg",
+	"version": "4.2.0",
+	"description": "orga processor for unifiedjs",
+	"files": [
+		"lib",
+		"index.js",
+		"index.d.ts",
+		"index.d.ts.map"
+	],
+	"type": "module",
+	"author": "Xiaoxing Hu <hi@xiaoxing.dev>",
+	"license": "MIT",
+	"homepage": "https://github.com/orgapp/orgajs/tree/main/packages/reorg#readme",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/orgapp/orgajs.git",
+		"directory": "packages/reorg"
+	},
+	"scripts": {},
+	"dependencies": {
+		"@orgajs/reorg-parse": "workspace:^",
+		"unified": "^11.0.5"
+	},
+	"devDependencies": {
+		"@types/unist": "^3.0.3"
+	}
 }
